+++ magicgui/backends/_qtpy/application.py
@@ -1,7 +1,7 @@
 from qtpy.QtCore import QTimer
 from qtpy.QtWidgets import QApplication
 
-from magicgui.protocols import BaseApplicationBackend
+from magicgui.widgets._protocols import BaseApplicationBackend
 
 
 # qt implementation
