+++ magicgui/types.py
@@ -1,4 +1,4 @@
-from enum import EnumMeta
+from enum import Enum, EnumMeta
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, Callable, Iterable, Optional, Tuple, Type, Union
 
@@ -6,8 +6,9 @@ from typing_extensions import TypedDict
 
 if TYPE_CHECKING:
     from magicgui.function_gui import FunctionGui
-    from magicgui.protocols import WidgetProtocol
-    from magicgui.widgets import CategoricalWidget, FileDialogMode, Widget
+    from magicgui.widgets._bases import CategoricalWidget, Widget
+    from magicgui.widgets._protocols import WidgetProtocol
+
 
 WidgetClass = Union[Type["Widget"], Type["WidgetProtocol"]]
 WidgetRef = Union[str, WidgetClass]
@@ -17,6 +18,21 @@ ChoicesCallback = Callable[["CategoricalWidget"], ChoicesIterable]
 ChoicesType = Union[EnumMeta, ChoicesIterable, ChoicesCallback, "ChoicesDict"]
 ReturnCallback = Callable[["FunctionGui", Any, Type], None]
 PathLike = Union[Path, str, bytes]
+
+
+class FileDialogMode(Enum):
+    """FileDialog mode options.
+
+    EXISTING_FILE - returns one existing file.
+    EXISTING_FILES - return one or more existing files.
+    OPTIONAL_FILE - return one file name that does not have to exist.
+    EXISTING_DIRECTORY - returns one existing directory.
+    """
+
+    EXISTING_FILE = "r"
+    EXISTING_FILES = "rm"
+    OPTIONAL_FILE = "w"
+    EXISTING_DIRECTORY = "d"
 
 
 class ChoicesDict(TypedDict):
@@ -37,4 +53,4 @@ class WidgetOptions(TypedDict, total=False):
     maximum: float
     step: float
     orientation: str
-    mode: Union[str, "FileDialogMode"]
+    mode: Union[str, FileDialogMode]
