+++ magicgui/widgets/__init__.py
@@ -1,0 +1,41 @@
+from ._concrete import (
+    CheckBox,
+    ComboBox,
+    Container,
+    DateTimeEdit,
+    FileEdit,
+    FloatSlider,
+    FloatSpinBox,
+    Label,
+    LineEdit,
+    LiteralEvalLineEdit,
+    LogSlider,
+    PushButton,
+    RadioButton,
+    RangeEdit,
+    SliceEdit,
+    Slider,
+    SpinBox,
+    TextEdit,
+)
+
+__all__ = [
+    "CheckBox",
+    "ComboBox",
+    "Container",
+    "DateTimeEdit",
+    "FileEdit",
+    "FloatSlider",
+    "FloatSpinBox",
+    "Label",
+    "LineEdit",
+    "LiteralEvalLineEdit",
+    "LogSlider",
+    "PushButton",
+    "RadioButton",
+    "RangeEdit",
+    "SliceEdit",
+    "Slider",
+    "SpinBox",
+    "TextEdit",
+]
